Add src/ dir and reorganize around modules
Leverage the modules added to move things down to a src/ dir. The logger
is now in a new core module. Also moves the utils/types/ to core/types/.
