+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "build": "tsc --project ./",
-    "test": "LOG_LEVEL=debug ts-mocha --paths --file ./tests/init \"**/*.spec*.ts\" --ignore \"node_modules/**/*\"",
+    "test": "LOG_LEVEL=debug ts-mocha --paths --file ./src/tests/init \"**/*.spec*.ts\" --ignore \"node_modules/**/*\"",
     "dev": "npm run build && LOG_LEVEL=debug node ./build/ ./sample-league.yaml | npx pino-pretty",
     "lint": "eslint --fix-dry-run .",
     "generate-docs": "node_modules/.bin/jsdoc -c jsdoc.json"
@@ -47,9 +47,10 @@
     "ts-mocha": "^8.0.0"
   },
   "_moduleAliases": {
-    "@formatting": "build/formatting",
-    "@generator": "build/generator",
-    "@models": "build/models",
-    "@utils": "build/utils"
+    "@core": "build/src/core",
+    "@formatting": "build/src/formatting",
+    "@generator": "build/src/generator",
+    "@models": "build/src/models",
+    "@utils": "build/src/utils"
   }
 }
