+++ src/core/types/either.spec.ts
