+++ src/core/types/pattern.ts
