+++ src/formatting/stringFormat.spec.ts
