+++ src/formatting/stringFormat.ts
