+++ tsconfig.json
@@ -43,8 +43,14 @@
 
     /* Module Resolution Options */
     // "moduleResolution": "node",            /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */
-    // "baseUrl": "./",                       /* Base directory to resolve non-absolute module names. */
-    // "paths": {},                           /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
+    "baseUrl": "./src/",                      /* Base directory to resolve non-absolute module names. */
+    "paths": {                                /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
+        "@core/*": ["core/*"],
+        "@formatting/*": ["formatting/*"],
+        "@generator/*": ["generator/*"],
+        "@models/*": ["models/*"],
+        "@utils/*": ["utils/*"]
+    },
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
     // "typeRoots": [],                       /* List of folders to include type definitions from. */
     // "types": [],                           /* Type declaration files to be included in compilation. */
@@ -66,12 +72,6 @@
     /* Advanced Options */
     "skipLibCheck": true,                     /* Skip type checking of declaration files. */
     "forceConsistentCasingInFileNames": true,  /* Disallow inconsistently-cased references to the same file. */
-    "resolveJsonModule": true,
-    "paths": {
-        "@formatting/*": ["formatting/*"],
-        "@generator/*": ["generator/*"],
-        "@models/*": ["models/*"],
-        "@utils/*": ["utils/*"]
-    }
+    "resolveJsonModule": true
   }
 }
